// server/src/generated/prisma/enums.ts
/* !!! This is code generated by Prisma. Do not edit directly. !!! */
/* eslint-disable */
// biome-ignore-all lint: generated file
// @ts-nocheck 
/*
* This file exports all enum related types from the schema.
*
* 🟢 You can import this file directly.
*/

export const Role = {
  ORGANIZER: 'ORGANIZER',
  CUSTOMER: 'CUSTOMER',
  GATE: 'GATE'
} as const

export type Role = (typeof Role)[keyof typeof Role]


export const EventType = {
  MOVIE: 'MOVIE',
  SHOW: 'SHOW'
} as const

export type EventType = (typeof EventType)[keyof typeof EventType]


export const EventSource = {
  MANUAL: 'MANUAL',
  TMDB: 'TMDB',
  TICKETMASTER: 'TICKETMASTER'
} as const

export type EventSource = (typeof EventSource)[keyof typeof EventSource]


export const EventStatus = {
  DRAFT: 'DRAFT',
  PUBLISHED: 'PUBLISHED',
  CANCELLED: 'CANCELLED'
} as const

export type EventStatus = (typeof EventStatus)[keyof typeof EventStatus]


export const ReservationStatus = {
  PENDING: 'PENDING',
  PAID: 'PAID',
  CANCELLED: 'CANCELLED',
  FAILED: 'FAILED'
} as const

export type ReservationStatus = (typeof ReservationStatus)[keyof typeof ReservationStatus]


export const TicketStatus = {
  VALID: 'VALID',
  USED: 'USED',
  CANCELLED: 'CANCELLED'
} as const

export type TicketStatus = (typeof TicketStatus)[keyof typeof TicketStatus]